requirements.txt
asgiref==3.2.3
astroid==2.3.3
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-appconf==1.0.3
django-ckeditor==5.9.0
django-debug-toolbar==2.2
django-imagekit==4.0.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-solo==1.1.3
djangorestframework==3.11.0
Faker==4.0.1
idna==2.8
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==7.0.0
pkg-resources==0.0.0
prompt-toolkit==3.0.5
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.9
wrapt==1.11.2
